refactor: split noun & pronoun from `DictWordMetadata` to `nominal.rs`
