+++ harper-core/src/lib.rs
@@ -44,9 +44,9 @@ pub use case::{Case, CaseIterExt};
 pub use char_string::{CharString, CharStringExt};
 pub use currency::Currency;
 pub use dict_word_metadata::{
-    AdverbData, ConjunctionData, Degree, DeterminerData, DictWordMetadata, NounData, PronounData,
-    dialect,
+    AdverbData, ConjunctionData, Degree, DeterminerData, DictWordMetadata, dialect,
     dialect::{Dialect, DialectFlags},
+    nominal::{NounData, PronounData},
     orthography::{OrthFlags, Orthography},
     verb::{VerbData, VerbForm, VerbFormFlags},
 };
